cynthion/requirements.txt
# - luna ----------------------------------------------------------------------

#luna @ git+https://github.com/greatscottgadgets/luna@main
#luna-soc @ git+https://github.com/greatscottgadgets/luna-soc@main

# - luna dev dependencies -----------------------------------------------------

# because: https://github.com/python-poetry/poetry/issues/3514

minerva @ git+https://github.com/lambdaconcept/minerva.git
amaranth-stdio @ git+https://github.com/amaranth-lang/amaranth-stdio.git@main
prompt-toolkit==3.0.16

# - forked dependencies -------------------------------------------------------

lambdasoc @ git+https://github.com/antoinevg/lambdasoc.git@antoinevg/support-vexriscv
greatfet @ git+https://github.com/antoinevg/greatfet@antoinevg/moondancer#subdirectory=host
pygreat @ git+https://github.com/antoinevg/libgreat.git@antoinevg/moondancer#subdirectory=host
facedancer @ git+https://github.com/antoinevg/Facedancer.git@antoinevg/moondancer
